Allow the binary option to be bubbled down from send to rawwebsocket's write_message.
It's not currently possible to send a message over the sockjs-tornado
"rawwebsocket" as binary instead of utf8.

Websocket defines both TextFrame and BinaryFrame when sending messages,
and there are different opcodes for each. This leads to issues with for
example when iOS websocket clients receives a TextFrame with binary in it.

Only the rawwebsocket's tranport has been updated to alter behavior when
the binary flag is set, as the other transports are javascript specific,
and javascript doesn't have support for binary.