small bugfix in multiplex example
The on_close method of MultiplexConnection was indented 
once too many times so it wasn't being called correctly. And the loop to 
close the endpoints once it was being called was trying to close the 
key of the dict vs the endpoint session obj. So fixed indent and 
updated close logic. 